requirements-dev.txt
-r requirements.txt
pytest==3.0.2
pytest-cov==2.3.1
pytest-django==3.0.0
pytest-sugar==0.8.0
flake8==3.0.4
codecov==2.0.9
requests_mock==1.3.0
